Lesson 24 - Environment variable/requirements.txt
python-dotenv==1.1.0
